--- app/_apis/schema/user.ts
@@ -1,10 +1,0 @@
-export interface SignUpRequest {
-  email: string;
-  nickName: string;
-  aboutMe: string;
-  job: string;
-}
-
-export interface SignUpResponse {
-  token: string;
-}
